feat: implement order processing and inventory management with new database schema and logic

+++ topics/transactions/database/migrations/007_product_inventory_table.sql
@@ -1,7 +1,7 @@
 CREATE TABLE inventory(
     id BIGINT GENERATED ALWAYS AS IDENTITY PRIMARY KEY,
     product_id BIGINT NOT NULL UNIQUE REFERENCES products(id) ON DELETE CASCADE,
-    stock INT NOT NULL check (stock>=0)
+    stock BIGINT NOT NULL check (stock>=0)
 )
 
 ---- create above / drop below ----
